SFML 3.0 migration for pioneer. Get the editor up and running.

+++ game/src/graphics/Background.cpp
@@ -48,9 +48,11 @@ void Background::render(automa::ServiceProvider& svc, sf::RenderWindow& win, sf:
 	for (auto& layer : layers) {
 		// backtrack sprites for infinite scroll effect
 		if (layer.physics.position.x < -scroll_pane.x && !locked_horizontally()) { layer.physics.position.x = 0.f; }
-		if (layer.physics.position.x > 0.f && !locked_horizontally()) { layer.physics.position.x = -scroll_pane.x; }
-		if (layer.physics.position.y < -scroll_pane.y && !locked_vertically()) { layer.physics.position.y = layer.physics.position.y + scroll_pane.y; }
-		if (layer.physics.position.y > 0.f && !locked_vertically()) { layer.physics.position.y = -scroll_pane.y + layer.physics.position.y; }
+		if (layer.physics.position.x > 0.f && !locked_horizontally()) { layer.physics.position.x = static_cast<float>(-scroll_pane.x); }
+		if (layer.physics.position.y < -scroll_pane.y && !locked_vertically()) { layer.physics.position.y = layer.physics.position.y + static_cast<float>(scroll_pane.y);
+		}
+		if (layer.physics.position.y > 0.f && !locked_vertically()) { layer.physics.position.y = static_cast<float>(-scroll_pane.y) + layer.physics.position.y;
+		}
 
 		layer.final_position = layer.physics.position - cam * layer.parallax;
